==> 02_Standardize_band_data.Rmd <==
---
title: "standarize band data"
author: "S. Sillen"
date: '2023-02-17'
output: html_document
---

# Required packages
```{r}
library(tidyverse)
library(lubridate)
library(colorscience)
library(broom)
library(feather)
```

Get correction coefficients (SR correction based on Gardner et al (2021) https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2020GL088946)
```{r}

data <- read_feather("C:/Users/samsi/OneDrive - University of Pittsburgh/LimnoSat/LimnoSat_20200628.feather") %>% 
  filter(across(c(Blue, Green, Red, Nir, Swir1, Swir2), ~ .x > 0 & .x < 10000)) %>% 
  mutate(year = year(date)) %>% 
  group_by(Hylak_id) %>%
  mutate(count =n(),
                  max_year=max(year, na.rm=T),
                  min_year = min(year, na.rm=T),
                  n_years = (max_year - min_year)) %>%
                  ungroup()

data$sat <- as.character(data$sat)

sr_57 <- data %>%
  filter(sat %in% c("l7", "l5")) %>%
  filter(date > "1999-01-01" & date < "2012-05-01" ) %>%
  # filter to site with enough data
  filter(n_years > 10) %>%
  select(Hylak_id, date, sat, count, n_years, Blue, Red, Green, Nir, Swir1, Swir2) %>%
  gather(Blue, Red, Green, Nir, Swir1, Swir2, key='band', value='value') 

# do ranking plotting percentiles, joining, and correcting
sr_57_rank  <- sr_57 %>%
  droplevels() %>%
  filter(sat =="l5") %>%
  group_by(band) %>%
  nest() %>%
  mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
          ret = purrr::invoke_map(tibble, ret)) %>%
  unnest(ret) %>%
  dplyr::select(-data) %>%
  pivot_longer(
    cols= contains("%")
  ) %>%
  mutate(quant = parse_number(name)/100) %>%
  rename(value_5 = value) %>%
  inner_join(sr_57 %>%
               droplevels() %>%
               filter(sat =="l7") %>%
               group_by(band) %>%
               nest() %>%
               mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
               ret = purrr::invoke_map(tibble, ret)) %>%
               unnest(ret) %>%
               dplyr::select(-data) %>%
               pivot_longer(
                 cols= contains("%")
               ) %>%
               mutate(quant = parse_number(name)/100) %>%
               rename(value_7 = value) %>%
               dplyr::select(-name),
             by=c("band", "quant")) 

poly_5_trunc <- function(df){
  lm(value_7 ~ poly(value_5, 2, raw=T), data = df %>%
       filter(!quant %in% c(0, 1))  )
}
poly_5_all <- function(df){
  lm(value_7 ~ poly(value_5, 2, raw=T), data = df)
}

## polynomial correction fit
poly_57 <- sr_57_rank %>%
  ungroup() %>%
#  filter(band != "dw") %>%
  nest(-band) %>%
  mutate( model = purrr::map(data, poly_5_trunc)) %>%
  mutate( model_all = purrr::map(data, poly_5_all)) %>%
  mutate( pred = purrr::map2(model, data, predict)) %>%
  mutate( pred_all = purrr::map2(model_all, data, predict)) %>%
  unnest(c(pred, pred_all, data))  %>%
  dplyr::select(-model, -model_all)

coef_5 <- sr_57_rank %>%
  ungroup() %>%
#  filter(band != "dw") %>%
  filter(!quant %in% c(0, 1)) %>%
  group_by(band)  %>%
  nest() %>%
  mutate( model = purrr::map(data, ~lm(value_7 ~ poly(value_5, 2, raw=T), data = .) %>%
                               tidy %>%
                               dplyr::select(term, estimate) %>%
                               spread(term, estimate))) %>%
  unnest(model) %>%
  dplyr::select(-data) %>%
  rename(band= 1, intercept=2, coef1=3, coef2=4 )  %>%
  mutate(sat = "5") %>%
  mutate(fit = "98_quant")

coef_5_all <- sr_57_rank %>%
  ungroup() %>%
  #  filter(band != "dw") %>%
  group_by(band)  %>%
  nest() %>%
  mutate( model = purrr::map(data, ~lm(value_7 ~ poly(value_5, 2, raw=T), data = .) %>%
                               tidy %>%
                               dplyr::select(term, estimate) %>%
                               spread(term, estimate))) %>%
  unnest(model) %>%
  dplyr::select(-data) %>%
  rename(band= 1, intercept=2, coef1=3, coef2=4 )  %>%
  mutate(sat = "5") %>%
  mutate(fit = "all_quant")


########################

sr_78 <- data %>%
  filter(sat %in% c("l7", "l8")) %>%
  filter(date > "2013-04-11" ) %>%
  # filter to site with enough data
  filter(n_years > 10) %>%
  select(Hylak_id, date, sat, count, n_years, Blue, Red, Green, Nir, Swir1, Swir2) %>%
  gather(Blue, Red, Green, Nir, Swir1, Swir2, key='band', value='value') 

# do ranking plotting percentiles, joining, and correcting
sr_78_rank  <- sr_78 %>%
  droplevels() %>%
  filter(sat =="l8") %>%
  group_by(band) %>%
  nest() %>%
  mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
          ret = purrr::invoke_map(tibble, ret)) %>%
  unnest(ret) %>%
  dplyr::select(-data) %>%
  pivot_longer(
    cols= contains("%")
  ) %>%
  mutate(quant = parse_number(name)/100) %>%
  rename(value_8 = value) %>%
  inner_join(sr_78 %>%
               droplevels() %>%
               filter(sat =="l7") %>%
               group_by(band) %>%
               nest() %>%
               mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
                       ret = purrr::invoke_map(tibble, ret)) %>%
               unnest(ret) %>%
               dplyr::select(-data) %>%
               pivot_longer(
                 cols= contains("%")
               ) %>%
               mutate(quant = parse_number(name)/100) %>%
               rename(value_7 = value) %>%
               dplyr::select(-name),
             by=c("band", "quant"))  

poly_8_trunc <- function(df){
  lm(value_7 ~ poly(value_8, 2), data = df %>%
       filter(!quant %in% c(0, 1))  )
}
poly_8_all <- function(df){
  lm(value_7 ~ poly(value_8, 2), data = df)
}

poly_78 <- sr_78_rank %>%
  ungroup() %>%
#  filter(band != "dw") %>%
  nest(-band) %>%
  mutate( model = purrr::map(data, poly_8_trunc)) %>%
  mutate( model_all = purrr::map(data, poly_8_all)) %>%
  mutate( pred = purrr::map2(model, data, predict)) %>%
  mutate( pred_all = purrr::map2(model_all, data, predict)) %>%
  unnest(c(pred, pred_all, data)) %>%
  dplyr::select(-model, -model_all)

coef_8 <- sr_78_rank %>%
  ungroup() %>%
  filter(!quant %in% c(0, 1)) %>%
  group_by(band)  %>%
  nest() %>%
  mutate( model = purrr::map(data, ~lm(value_7 ~ poly(value_8, 2, raw=T), data = .) %>%
                               tidy %>%
                               dplyr::select(term, estimate) %>%
                               spread(term, estimate))) %>%
  unnest(model) %>%
  dplyr::select(-data) %>%
  rename(band= 1, intercept=2, coef1=3, coef2=4 )  %>%
  mutate(sat = "8") %>%
  mutate(fit = "98_quant")

coef_8_all <- sr_78_rank %>%
  ungroup() %>%
  group_by(band)  %>%
  nest() %>%
  mutate( model = purrr::map(data, ~lm(value_7 ~ poly(value_8, 2, raw=T), data = .) %>%
                               tidy %>%
                               dplyr::select(term, estimate) %>%
                               spread(term, estimate))) %>%
  unnest(model) %>%
  dplyr::select(-data) %>%
  rename(band= 1, intercept=2, coef1=3, coef2=4 )  %>%
  mutate(sat = "8") %>%
  mutate(fit = "all_quant")


coef_7 <- tibble(band = c("Blue", "Red", "Green", "Nir", "Swir1", "Swir2"), intercept = 0, coef1=1, coef2=0, sat= "7")


corr_coef <- bind_rows(coef_5, coef_7, coef_8, coef_5_all, coef_8_all) %>%
  ungroup()

write.csv(corr_coef, "C:/Users/samsi/OneDrive - University of Pittsburgh/lake_productivity_trends/Data/corr_coef.csv", row.names = FALSE)
```  

Add corrected band values to training set 
```{r}
training <- read_csv("C:/Users/samsi/OneDrive - University of Pittsburgh/lake_productivity_trends/Data/training_final.csv") 

correction_coef_99 <- corr_coef %>%
  filter(fit %in% c(NA, "98_quant")) %>%
  dplyr::select(-fit) %>%
  mutate(band = tolower(band))

training <- training %>%
  mutate(rn = row_number()) %>%
  mutate(sat = as.character(sat)) %>%
  gather(red ,green, blue, nir, swir1, swir2, key='band', value='value') %>%
  group_by(band, sat) %>%
  left_join(correction_coef_99, by=c("band", "sat")) %>%
  mutate(value_cor=  coef2*value^ 2 + coef1*value + intercept) %>%
  ungroup() %>%
  mutate(value_cor = ifelse(value_cor <=0, value, value_cor)) %>%
  dplyr::select(-intercept, -coef1, -coef2) %>%
  pivot_wider(names_from = band,
              values_from = c("value", "value_cor"))  %>%
    rename_at(vars(starts_with("value_")),
           function(x) stringr::str_replace_all(x, "value_", "")) %>%
    rename_at(vars(red, green, blue, nir, swir1, swir2),function(x) paste0(x,"_raw")) %>%
    rename_at(vars(starts_with("cor")),            funs(stringr::str_replace_all(., "cor_", ""))) 

write.csv(training, "C:/Users/samsi/OneDrive - University of Pittsburgh/lake_productivity_trends/Data/training_standardized.csv", row.names = FALSE)
```
