==> 06_Driver_analysis.rmd <==
---
title: "06_Driver_analysis"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Required packages 
```{r}
library(USAboundaries)
library(tidyverse)
library(ggpubr)
library(sf)
library(grid)
library(feather)
library(randomForest)
library(caret)
library(vip)
library(corrplot)
library(leafgl)
library(raster)
```

# Make trend map (Figure 4)
```{r}
states <- us_states()

roi <- states %>% filter(state_abbr %in% c('CA','NE','KS','ID', 'MT', 'UT', 'CO', 'WY', 'SD', 'ND', 'NV', 'WA', 'OR', 'AZ'))

grob1 <- grobTree(textGrob("WYOMING", x=0.67,  y=0.54, hjust=0,
                          gp=gpar(col="#8a8a8a", fontsize=13)))
grob2 <- grobTree(textGrob("MONTANA", x=0.57,  y=0.73, hjust=0,
                           gp=gpar(col="#8a8a8a", fontsize=13)))
grob3 <- grobTree(textGrob("IDAHO", x=0.21,  y=0.52, hjust=0,
                           gp=gpar(col="#8a8a8a", fontsize=13)))
grob4 <- grobTree(textGrob("UTAH", x=0.42,  y=0.16, hjust=0,
                           gp=gpar(col="#8a8a8a", fontsize=13)))
grob5 <- grobTree(textGrob("COLORADO", x=0.82,  y=0.2, hjust=0,
                           gp=gpar(col="#8a8a8a", fontsize=13)))

# get location data 
lakes <- read_sf("/Users/samsillen/OneDrive - University of Pittsburgh/LimnoSat/HydroLakes_DP.shp") %>%
filter(type == 'dp') 

# load in trend data , join to location data
trends <- read_csv("/Users/samsillen/OneDrive - University of Pittsburgh/WRR_final/Data/trends.csv")%>% 
  
left_join(lakes) %>% 
st_as_sf()

pal <- colorFactor(c( '#2158bc',  '#698c86', '#7dae38'), domain = c("improving", "no change", "worsening" ))

trends$trend <- factor(trends$trend, levels = c('no trend', 'trending oligotrophic', 'trending eutrophic'))


ggplot()+
  geom_sf(data = roi, fill = "#292b30") +
  geom_sf(data = trends, size = 2.5, alpha = 0.9,aes(fill = trend), pch = 21, colour = 'black')+
  scale_fill_manual(values=c( '#698c86', '#2158bc', '#7dae38')) +
  coord_sf(
    ylim = c(36.99, 49.8),
    xlim = c(-117.5,-101),
    clip = "on",
    expand = F
  ) +
  guides(fill = guide_legend(override.aes = list(size=5)))+
  labs(x = "Longitude", y = "Latitude") +
  theme_bw() +
  theme(legend.title = element_blank(), legend.position = "bottom", axis.title = element_text(size = 20),panel.border  = element_rect(colour = "white"), panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.text = element_text(size = 16))
  
```

# Driver analysis 
```{r}
# load in explanatory variable data 
var_data <- read_feather("/Users/samsillen/OneDrive - University of Pittsburgh/WRR_final/Data/preds.feather") %>% 
select(-type, -distance) %>%
filter(Hylak_id %in% trends$Hylak_id) %>% 
distinct(Hylak_id, .keep_all = TRUE) %>% 
left_join(trends) %>%
  mutate(Development = PctUrbHi2006Cat + PctUrbLo2006Cat + PctUrbMd2006Cat, 
         Forest = PctMxFst2006Cat + PctDecid2006Cat + PctConif2006Cat, 
         Wetland = PctWdWet2006Cat + PctHbWet2006Cat )

# add in hydrolakes data 
Hydrolakes <- read_sf("/Users/samsillen/OneDrive - University of Pittsburgh/WesternMountainsChlPreds/HydroLAKES_polys_v10_shp/HydroLAKES_polys_v10.shp") %>% 
  dplyr::filter(Hylak_id %in% var_data$Hylak_id) %>% 
  dplyr::select(Hylak_id, Depth_avg, Slope_100, Res_time, Vol_total, Lake_area)

var_data <- left_join(var_data, Hydrolakes, by = 'Hylak_id')
```

```{r}
predictors <- c("lake_elevation_m",  "Tmean8110Cat",      "ClayCat", 
"BFICat",           "CatAreaSqKm",       "PctAg2006Slp10Cat",     "HydrlCondCat",    
"Wetland",
"KffactCat",         "NO3_2008Cat",       "RunoffCat", 
"WtDepCat", "Forest", "Development", "Depth_avg", "Slope_100", "Res_time", "Vol_total", "Lake_area")
response <- "trend"
ctrl_default <- trainControl(method="cv", number = 3, savePredictions = TRUE)

set.seed(100)

var_data <- var_data %>% as.data.frame() 
#run on 15, then keep trying 
val_ids <- var_data  %>% group_by(trend) %>% slice_sample(prop = 0.15)

## Set up our data partitoins
train_d = var_data %>% filter(!lagoslakeid %in% val_ids$lagoslakeid) 
test_d = var_data %>% filter(lagoslakeid %in% val_ids$lagoslakeid) 

set.seed(100)

trControl <- trainControl(method = "cv",
                          number = 10,
                          search = "grid")

rf_default <- train(x = train_d[,predictors],
                    y = train_d[,response],
                    method = "rf",
                    metric = "Accuracy",
                    trControl = trControl)
# Print the results
print(rf_default)
#Step 2) Search best mtry
tuneGrid <- expand.grid(.mtry = c(10: 20))
rf_mtry <- train(x = train_d[,predictors],
                 y = train_d[,response],
                 method = "rf",
                 metric = "Accuracy",
                 tuneGrid = tuneGrid,
                 trControl = trControl,
                 importance = TRUE,
                 nodesize = 14,
                 ntree = 300)
print(rf_mtry)
rf_mtry$bestTune$mtry #best value
best_mtry <- rf_mtry$bestTune$mtry 
best_mtry
#3) search the best maxnodes
store_maxnode <- list()
tuneGrid <- expand.grid(.mtry = best_mtry)
for (maxnodes in c(5: 20)) {
  set.seed(1234)
  rf_maxnode <- train(x = train_d[,predictors],
                      y = train_d[,response],
                      method = "rf",
                      metric = "Accuracy",
                      tuneGrid = tuneGrid,
                      trControl = trControl,
                      importance = TRUE,
                      nodesize = 14,
                      maxnodes = maxnodes,
                      ntree = 300)
  current_iteration <- toString(maxnodes)
  store_maxnode[[current_iteration]] <- rf_maxnode
}
results_mtry <- resamples(store_maxnode)
summary(results_mtry)
#Step 4) Search the best ntrees
store_maxtrees <- list()
for (ntree in c(250, 300, 350, 400, 450, 500, 550, 600, 800, 1000, 2000)) {
  set.seed(5678)
  rf_maxtrees <- train(x = train_d[,predictors],
                       y = train_d[,response],                       method = "rf",
                       metric = "Accuracy",
                       tuneGrid = tuneGrid,
                       trControl = trControl,
                       importance = TRUE,
                       nodesize = 10,
                       maxnodes = 7,
                       ntree = ntree)
  key <- toString(ntree)
  store_maxtrees[[key]] <- rf_maxtrees
}
results_tree <- resamples(store_maxtrees)
summary(results_tree)
fit_rf <- randomForest(x = train_d[,predictors],
                       y = train_d[,response],       
                       method = "rf",
                       metric = "Accuracy",
                       tuneGrid = tuneGrid,
                       trControl = trControl,
                       importance = TRUE,
                       nodesize = 10,
                       ntree = 300,
                       maxnodes = 7)
prediction <-predict(fit_rf, test_d)
confMatRF<-confusionMatrix(prediction, test_d$trend)
varImpPlot(fit_rf)
imp_df<-importance(fit_rf)
imp_df
vip(fit_rf, geom = "col", horizontal = TRUE, size = 1.5, plot.engine="ggplot2", num_features = 5)

```

# This chunk makes figure 5 
```{r}
Figure5 <- var_data %>% 
  dplyr::select(Res_time, Forest, Tmean8110Cat, BFICat, RunoffCat, trend) 

trendColors <- c('no trend' =  '#698c86', 'trending oligotrophic' = '#2158bc', 'trending eutrophic' = '#7dae38')

Restime <- Figure5 %>% 
dplyr::select(trend, Res_time) %>% 
group_by(trend) %>% 
mutate(Res_time_filtered = case_when(Res_time - quantile(Res_time)[4] > 1.5*IQR(Res_time) ~ NA_real_,
                                  quantile(Res_time)[2] - Res_time > 1.5*IQR(Res_time) ~ NA_real_,
                                  TRUE ~ Res_time)) 

Restime_plot <- ggplot(Restime, aes(x = trend, y = Res_time_filtered)) +
  #geom_point() +
  #geom_pointrange(aes(ymin = quant25, ymax = quant75)) +
  geom_boxplot(outlier.shape = NA) +
    geom_jitter(
      aes(fill = trend),
      shape = 21,
      size = 1.5,
      alpha = 0.6,
      position = position_jitter(0.2)
    ) +  
  ylab("Residence Time") + 
  ylim(0, 10000) +
  scale_fill_manual(values = trendColors) + 
theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), axis.title.x  = element_blank(), 
        legend.position = 'none', legend.text = element_text(face = 'bold'), axis.title.y = element_text(size = 14, face = 'bold'), axis.text.x = element_blank(), axis.title = element_text(size = 16))

Tmean <- Figure5 %>% 
dplyr::select(trend, Tmean8110Cat) %>% 
group_by(trend) %>% 
mutate(Tmean_filtered = case_when(Tmean8110Cat - quantile(Tmean8110Cat)[4] > 1.5*IQR(Tmean8110Cat) ~ NA_real_,
                                  quantile(Tmean8110Cat)[2] - Tmean8110Cat > 1.5*IQR(Tmean8110Cat) ~ NA_real_,
                                  TRUE ~ Tmean8110Cat))

Tmean_plot <- ggplot(Tmean, aes(x = trend, y = Tmean_filtered)) +
  #geom_point() +
  #geom_pointrange(aes(ymin = quant25, ymax = quant75)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(
    aes(fill = trend),
    shape = 21,
    size = 1.5,
    alpha = 0.6,
    position = position_jitter(0.2)
  ) +  
  ylab("30 Year Mean Temp") +
  scale_fill_manual(values = trendColors) + 
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), axis.title.x  = element_blank(), 
        legend.position = 'none', legend.text = element_text(face = 'bold'), axis.title.y = element_text(size = 14, face = 'bold'), axis.text.x = element_blank(), axis.title = element_text(size = 16))


Forest <- Figure5 %>% 
dplyr::select(trend, Forest) %>% 
group_by(trend) %>% 
mutate(Forest_filtered = case_when(Forest - quantile(Forest)[4] > 1.5*IQR(Forest) ~ NA_real_,
                                  quantile(Forest)[2] - Forest > 1.5*IQR(Forest) ~ NA_real_,
                                  TRUE ~ Forest))



Forest_plot <- ggplot(Forest, aes(x = trend, y = Forest_filtered)) +
  #geom_point() +
  #geom_pointrange(aes(ymin = quant25, ymax = quant75)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(
    aes(fill = trend),
    shape = 21,
    size = 1.5,
    alpha = 0.6,
    position = position_jitter(0.2)
  ) +  
  ylab("% Forest") +
  scale_fill_manual(values = trendColors) + 
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), axis.title.x  = element_blank(), 
        legend.position = 'none', legend.text = element_text(face = 'bold'), axis.title.y = element_text(size = 14, face = 'bold'), axis.text.x = element_blank(), axis.title = element_text(size = 16))


BFICat <- Figure5 %>% 
dplyr::select(trend, BFICat) %>% 
group_by(trend) %>% 
mutate(BFI_filtered = case_when(BFICat - quantile(BFICat)[4] > 1.5*IQR(BFICat) ~ NA_real_,
                                  quantile(BFICat)[2] - BFICat > 1.5*IQR(BFICat) ~ NA_real_,
                                  TRUE ~ BFICat))


BFI_plot <- ggplot(BFICat, aes(x = trend, y = BFI_filtered)) +
  #geom_point() +
  #geom_pointrange(aes(ymin = quant25, ymax = quant75)) +
  geom_boxplot(outlier.shape = NA) + 
  geom_jitter(
    aes(fill = trend),
    shape = 21,
    size = 1.5,
    alpha = 0.6,
    position = position_jitter(0.2)
  ) +  
  ylab( "BFI") +
  scale_fill_manual(values = trendColors) + 
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), axis.title.x  = element_blank(), 
        legend.position = 'none', legend.text = element_text(face = 'bold'), axis.title.y = element_text(size = 14, face = 'bold'), axis.text.x = element_blank(), axis.title = element_text(size = 16))

Runoff <- Figure5 %>% 
dplyr::select(trend, RunoffCat) %>% 
group_by(trend) %>% 
mutate(Runoff_filtered = case_when(RunoffCat - quantile(RunoffCat)[4] > 1.5*IQR(RunoffCat) ~ NA_real_,
                                  quantile(RunoffCat)[2] - RunoffCat > 1.5*IQR(RunoffCat) ~ NA_real_,
                                  TRUE ~ RunoffCat))

Runoff_plot <- ggplot(Runoff, aes(x = trend, y = Runoff_filtered)) +
  #geom_point() +
  #geom_pointrange(aes(ymin = quant25, ymax = quant75)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(
    aes(fill = trend),
    shape = 21,
    size = 1.5,
    alpha = 0.6,
    position = position_jitter(0.2)
  ) +  
  ylab( "Runoff") +
  scale_fill_manual(values = trendColors) + 
    guides(fill = guide_legend(override.aes = list(size=5)))+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), axis.title.x  = element_blank(), 
        legend.position = 'bottom', legend.title = element_blank(),legend.text = element_text(size = 14), axis.title.y = element_text(size = 14, face = 'bold'), axis.text.x = element_blank(), axis.title = element_text(size = 16))

#arrange the plots
ggarrange(Restime_plot, Forest_plot, Tmean_plot, BFI_plot, Runoff_plot, nrow = 5, heights = c(5,5), align = "v")

```

# This chunk makes figure 6
```{r}
myDF <- data.frame(variables = row.names(imp_df), imp_df)
renamer <- tibble(variables = c(
  "lake_elevation_m",  "Tmean8110Cat",      "ClayCat", 
"BFICat",  "CatAreaSqKm",       "PctAg2006Slp10Cat",     "HydrlCondCat",    
"Wetland",
"KffactCat", "NO3_2008Cat",       "RunoffCat", 
"WtDepCat", "Forest", "Development", "Depth_avg", "Slope_100", "Res_time", "Vol_total", "Lake_area"),
  
  namesNew = c(
    "Lake Elevation",  "30 Year Mean Temp", "% Clay Content", "Base Flow Index (% of Total Flow)",  "Area of Catchment (Sq Km)","% Agricultural Land on Slopes > 20%", 
    "Hydraulic Conductivity",  "% Wetland",    "Kffactor",         "Mean Wet Deposition of Nitrate (kg/ha/yr)",       "Runoff", "Water Table Depth",      
    "% Forest",    "% Development", "Lake Depth","Slope", "Residence Time",  "Total Volume (million cubic meters)",  "Lake Area" )) 

myDF <- left_join(myDF, renamer)

no_trend <- myDF %>% select(variables, no.trend, namesNew) %>% 
  filter(variables %in% c('Tmean8110Cat', 'HydrlCondCat', 'Slope_100', 'Forest','BFICat')) %>% 
  mutate(names2 = fct_reorder(namesNew, no.trend))

no_trend_var_imp <- ggplot(no_trend) +
  geom_col(aes(y = names2, x = no.trend), width = 0.5) +
  theme_bw() + 
  xlab("Mean Decrease in Accuracy") + 
  labs(title = "a) No Trend")+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.title = element_text(face = "bold"), axis.text.x = element_text(size = 12), axis.text.y = element_text(size = 14, colour = 'black'), axis.title.x = element_text(size = 16, 'black'), axis.title.y = element_blank(), title = element_text(size = 16))

inc_oligotrophic <- myDF %>% select(variables, trending.oligotrophic, namesNew) %>% 
  filter(variables %in% c('Res_time', 'ClayCat', 'Development', 'lake_elevation_m','Forest')) %>% 
  mutate(names2 = fct_reorder(namesNew, trending.oligotrophic))

inc_oligotrophic_var_imp <- ggplot(inc_oligotrophic) +
  geom_col(aes(y = names2, x = trending.oligotrophic), width = 0.5) +
  theme_bw() + 
  xlab("Mean Decrease in Accuracy") + 
  labs(title = "b) Trending Oligotrophic") + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.title = element_text(face = "bold"), axis.text.x = element_text(size = 12, colour = 'black'), axis.text.y = element_text(size = 14, colour = 'black'), axis.title.x = element_text(size = 16), axis.title.y = element_blank(), title = element_text(size = 16))

inc_eutrophic <- myDF %>% select(variables, trending.eutrophic, namesNew)%>% 
  filter(variables %in% c('PctAg2006Slp10Cat', 'CatAreaSqKm', 'NO3_2008Cat', 'Wetland','Forest')) %>% 
  mutate(names2 = fct_reorder(namesNew, trending.eutrophic))

inc_eutrophic_var_imp <- ggplot(inc_eutrophic) +
  geom_col(aes(y = names2, x = trending.eutrophic), width = 0.5) +
  theme_bw() + 
  xlab("Mean Decrease in Accuracy") + 
  labs(title = "c) Trending Eutrophic") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.title = element_text(face = "bold"), axis.text.x = element_text(size = 12, colour = 'black'), axis.text.y = element_text(size = 14, colour = 'black'), axis.title.x = element_text(size = 16), axis.title.y = element_blank(), title = element_text(size = 16))

#Figure 6
ggarrange(no_trend_var_imp, inc_oligotrophic_var_imp, inc_eutrophic_var_imp, nrow = 3, heights = c(4, 4), align = "v")
```

# Make correlation matrix for supplementary material 
```{r}
corr_data <- var_data %>% 
select("lake_elevation_m",  "Tmean8110Cat",      "ClayCat", 
"BFICat",  "CatAreaSqKm",       "PctAg2006Slp10Cat",     "HydrlCondCat",    
"Wetland",
"KffactCat", "NO3_2008Cat",       "RunoffCat", 
"WtDepCat", "Forest", "Development", "Depth_avg", "Slope_100", "Res_time", "Vol_total", "Lake_area") 

corr_data <- corr_data %>% 
rename('Lake Elevation' = lake_elevation_m, '30 Year Mean Temp' =  Tmean8110Cat,
 '% Clay Content' =      ClayCat, 
'Base Flow Index ' = BFICat, 'Area of Catchment ' = CatAreaSqKm,
 '% Agricultural Land on Slopes > 20%' =      PctAg2006Slp10Cat,
   'Hydraulic Conductivity' =  HydrlCondCat,  '% Wetland' = Wetland,
'Kffactor' = KffactCat, 'Mean Wet Deposition of Nitrate' = NO3_2008Cat,  
    'Runoff' =  RunoffCat, 'Water Table Depth' = WtDepCat, '% Forest' = Forest, 
    '% Development' = Development, 'Lake Depth' = Depth_avg, 'Slope' = Slope_100,
     'Residence Time' =  Vol_total, 'Lake Area' = Lake_area)

mydata.corr = cor(corr_data, method = c("spearman"))

corrplot(mydata.corr, type = 'lower',  tl.col = 'black', diag = FALSE, tl.cex = 1.2)
```
